hum je sais plus ce que je faisais

+++ array.hpp
@@ -31,8 +31,8 @@ class Array {
 		Iterator &operator++() { this->idx++; return *this; }
 		Iterator &operator--() { this->idx--; return *this; }
 		
-		Iterator &operator+(size_t val) { return Iterator(this->start, this->idx + val); }
-		Iterator &operator-(size_t val) { return Iterator(this->start, this->idx - val); }
+		Iterator &operator+(int val) { return Iterator(this->start, this->idx + val); }
+		Iterator &operator-(int val) { return Iterator(this->start, this->idx - val); }
 
 		auto operator<=>(const Iterator& other) const { return this->idx <=> other.idx; }
 		bool operator==(const Iterator& other) const { return this->idx == other.idx; }
